Move things into coe implementation module
