+++ core-android/build.gradle
@@ -19,22 +19,9 @@ android {
 }
 
 dependencies {
-  api project(':core-api')
-  implementation project(':core')
-
   api 'androidx.appcompat:appcompat:1.2.0'
   api 'androidx.annotation:annotation:1.1.0'
   api 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-  api 'androidx.lifecycle:lifecycle-reactivestreams-ktx:2.2.0'
-
-  kapt rootProject.ext.daggerAnnotationProcessor
-  implementation rootProject.ext.dagger
-
-  implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
   api 'com.jakewharton.timber:timber:4.7.1'
-
-  testImplementation 'junit:junit:4.13'
-  testImplementation 'org.assertj:assertj-core:3.16.1'
-  testImplementation 'org.mockito:mockito-core:3.5.6'
 }
