+++ core/build.gradle
@@ -1,18 +1,37 @@
-apply plugin: 'java-library'
-apply plugin: 'kotlin'
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+
+android {
+  compileSdkVersion 29
+  defaultConfig {
+    minSdkVersion 21
+  }
+
+  compileOptions {
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
+  }
+
+  kotlinOptions {
+    jvmTarget = JavaVersion.VERSION_1_8.toString()
+  }
+}
 
 dependencies {
   implementation project(':core-api')
+  implementation project(':core-android')
 
-  implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+  implementation 'androidx.appcompat:appcompat:1.2.0'
+  implementation 'androidx.annotation:annotation:1.1.0'
+  implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 
-  api 'io.reactivex.rxjava2:rxjava:2.2.19'
-  api rootProject.ext.okHttp
-  api 'com.squareup.okio:okio:2.7.0'
-  implementation rootProject.ext.retrofit
+  kapt rootProject.ext.daggerAnnotationProcessor
+  implementation rootProject.ext.dagger
 
-  api 'org.threeten:threetenbp:1.4.4:no-tzdb'
-  api 'javax.inject:javax.inject:1'
+  implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+
+  implementation 'com.jakewharton.timber:timber:4.7.1'
 
   testImplementation 'junit:junit:4.13'
   testImplementation 'org.assertj:assertj-core:3.16.1'
