+++ core/src/main/AndroidManifest.xml
@@ -1,0 +1,19 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="com.jraska.github.client.core">
+
+  <application>
+    <activity android:name=".android.UriHandlerActivity">
+      <intent-filter>
+        <action android:name="android.intent.action.VIEW"/>
+
+        <category android:name="android.intent.category.DEFAULT"/>
+        <category android:name="android.intent.category.BROWSABLE"/>
+
+        <data
+          android:host="github.com"
+          android:pathPattern="/.*"
+          android:scheme="https"/>
+      </intent-filter>
+    </activity>
+  </application>
+</manifest>
