+++ core/src/main/java/com/jraska/github/client/core/android/UriHandlerViewModel.kt
