+++ core/src/main/java/com/jraska/github/client/core/android/logging/ErrorReportTree.kt
